refactor(editor): Move NodeView template operations on useCanvasOperations for reusability
